Simplify the `EncoderChooser` a bit more
