--- pkg/preparation/dags/visitor/unixfsencoder.go
@@ -1,40 +1,0 @@
-package visitor
-
-import (
-	"fmt"
-	"io"
-
-	"github.com/ipfs/go-cid"
-	dagpb "github.com/ipld/go-codec-dagpb"
-	"github.com/ipld/go-ipld-prime/codec"
-	"github.com/ipld/go-ipld-prime/datamodel"
-)
-
-type unixFSEncoder struct {
-	visitUnixFSNode func(cid cid.Cid, size uint64, ufsData []byte, pbLinks []dagpb.PBLink, data []byte) error
-	originalEncode  codec.Encoder
-}
-
-// Encode encodes the node as a UnixFS node and visits it using the
-// UnixFSNodeVisitor. The node must be a PBNode, containing UnixFS data.
-func (e unixFSEncoder) Encode(node datamodel.Node, w io.Writer) error {
-	cid, data, err := encode(e.originalEncode, cid.DagProtobuf, node, w)
-	if err != nil {
-		return fmt.Errorf("encoding node: %w", err)
-	}
-	pbNode, ok := node.(dagpb.PBNode)
-	if !ok {
-		return fmt.Errorf("failed to cast node to PBNode")
-	}
-	ufsData := pbNode.FieldData().Must().Bytes()
-	links := make([]dagpb.PBLink, 0, pbNode.FieldLinks().Length())
-	iter := pbNode.FieldLinks().Iterator()
-	for !iter.Done() {
-		_, link := iter.Next()
-		links = append(links, link)
-	}
-	if err := e.visitUnixFSNode(cid, uint64(len(data)), ufsData, links, data); err != nil {
-		return fmt.Errorf("visiting unixfs node: %w", err)
-	}
-	return nil
-}
